Django/requirements.txt
altair==5.5.0
asgiref==3.8.1
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
click==8.1.8
Django==5.0.14
django-cors-headers==4.7.0
djangorestframework==3.16.0
gitdb==4.0.12
GitPython==3.1.44
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mssql-django==1.5
narwhals==1.34.0
protobuf==5.29.4
pyarrow==19.0.1
pydeck==0.9.1
pyodbc==5.2.0
pytz==2025.2
referencing==0.36.2
rpds-py==0.24.0
smmap==5.0.2
sqlparse==0.5.3
streamlit==1.44.1
tenacity==9.1.2
toml==0.10.2
tornado==6.4.2
tzdata==2025.2
watchdog==6.0.0
